🤖 Refactor(ncloud-sdk): server와 vserver를 vserver로 통합
