--- packages/ncloud-sdk/src/services/subnet/signature.ts
@@ -1,36 +1,0 @@
-import { createHmac } from 'crypto';
-import { ApiKeyCredentials, RequestConfig } from './types';
-
-interface SignatureConfig extends RequestConfig {
-    apiKey: ApiKeyCredentials;
-}
-
-export function generateSignature({
-    method,
-    url,
-    timestamp,
-    params,
-    apiKey,
-}: SignatureConfig): string {
-    const space = ' ';
-    const newLine = '\n';
-
-    const fullUrl = params
-        ? `${url}?${new URLSearchParams(params).toString()}`
-        : url;
-
-    console.log(fullUrl);
-    const message = [
-        method,
-        space,
-        fullUrl,
-        newLine,
-        timestamp,
-        newLine,
-        apiKey.accessKey,
-    ].join('');
-
-    return createHmac('sha256', apiKey.secretKey)
-        .update(message)
-        .digest('base64');
-}
